move interests into components
