+++ src/Components/App/App.js
@@ -8,7 +8,7 @@ import NotFound from '../NotFound';
 import PlanTrip from '../PlanTrip/PlanTrip';
 import LandingPage from '../LandingRoute/LandingPage';
 import MapContainer from '../MapContainer/MapContainer';
-import Interests from '../../Interests/Interests';
+import Interests from '../Interests/Interests';
 import { ContextProvider } from '../../Context';
 import WaypointsSelect from '../waypointsSelect/WaypointsSelect'
 import './App.css';
